Rename component to common
