+++ ui/src/application/Applications.tsx
@@ -10,9 +10,9 @@ import TableRow from '@material-ui/core/TableRow';
 import Delete from '@material-ui/icons/Delete';
 import Edit from '@material-ui/icons/Edit';
 import React, {ChangeEvent, Component, SFC} from 'react';
-import ConfirmDialog from '../component/ConfirmDialog';
-import DefaultPage from '../component/DefaultPage';
-import ToggleVisibility from '../component/ToggleVisibility';
+import ConfirmDialog from '../common/ConfirmDialog';
+import DefaultPage from '../common/DefaultPage';
+import ToggleVisibility from '../common/ToggleVisibility';
 import AddApplicationDialog from './AddApplicationDialog';
 import {observer} from 'mobx-react';
 import {observable} from 'mobx';
