+++ ui/src/client/Clients.tsx
@@ -8,9 +8,9 @@ import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
 import Delete from '@material-ui/icons/Delete';
 import React, {Component, SFC} from 'react';
-import ConfirmDialog from '../component/ConfirmDialog';
-import DefaultPage from '../component/DefaultPage';
-import ToggleVisibility from '../component/ToggleVisibility';
+import ConfirmDialog from '../common/ConfirmDialog';
+import DefaultPage from '../common/DefaultPage';
+import ToggleVisibility from '../common/ToggleVisibility';
 import AddClientDialog from './AddClientDialog';
 import {observer} from 'mobx-react';
 import {observable} from 'mobx';
