+++ ui/src/layout/Layout.tsx
@@ -4,10 +4,10 @@ import axios, {AxiosResponse} from 'axios';
 import * as React from 'react';
 import {HashRouter, Redirect, Route, Switch} from 'react-router-dom';
 import Header from './Header';
-import LoadingSpinner from '../component/LoadingSpinner';
+import LoadingSpinner from '../common/LoadingSpinner';
 import Navigation from './Navigation';
-import ScrollUpButton from '../component/ScrollUpButton';
-import SettingsDialog from '../component/SettingsDialog';
+import ScrollUpButton from '../common/ScrollUpButton';
+import SettingsDialog from '../common/SettingsDialog';
 import SnackBarHandler from '../snack/SnackBarHandler';
 import * as config from '../config';
 import Applications from '../application/Applications';
