+++ ui/src/pages/Login.tsx
@@ -2,8 +2,8 @@ import Button from '@material-ui/core/Button';
 import Grid from '@material-ui/core/Grid';
 import TextField from '@material-ui/core/TextField';
 import React, {Component, FormEvent} from 'react';
-import Container from '../component/Container';
-import DefaultPage from '../component/DefaultPage';
+import Container from '../common/Container';
+import DefaultPage from '../common/DefaultPage';
 import {observable} from 'mobx';
 import {observer} from 'mobx-react';
 import {inject, Stores} from '../inject';
