+++ ui/src/user/Users.tsx
@@ -10,8 +10,8 @@ import TableRow from '@material-ui/core/TableRow';
 import Delete from '@material-ui/icons/Delete';
 import Edit from '@material-ui/icons/Edit';
 import React, {Component, SFC} from 'react';
-import ConfirmDialog from '../component/ConfirmDialog';
-import DefaultPage from '../component/DefaultPage';
+import ConfirmDialog from '../common/ConfirmDialog';
+import DefaultPage from '../common/DefaultPage';
 import AddEditDialog from './AddEditUserDialog';
 import {observer} from 'mobx-react';
 import {observable} from 'mobx';
